ImageGalleriComponent.vue virker og jeg har fixet en masse andre små fejl, så siden begynder at se helt ok ud.
